requirements.txt
firecrawl-py>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio>=3.4.3